requirements.txt
pynvim==0.5.0
llama-index-llms-ollama==0.1.2
